requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==4.2.16
django-cors-headers==4.6.0
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
python-dotenv==1.0.1
pydantic==2.11.5
redis==5.2.0
sqlparse==0.5.2
typing_extensions==4.12.2
psycopg2==2.9.9
celery==5.4.0
setuptools==80.4.0
pillow==10.4
crawl4ai==0.6.3
openai==1.75.0
docling==2.30.0
ollama==0.4.7
chromadb==1.0.10
langchain==0.3.14
langchain-chroma==0.2.4
langchain-community==0.3.14
langchain-core==0.3.60
langchain-docling==0.2.0
langchain-ollama==0.2.3
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langgraph==0.2.69